requirements.txt
tensorflow
gradio
numpy
Pillow
